requirements.txt
streamlit == 0.82.0
keras==2.3.1
pillow ==7.0
numpy >=1.9.2
gunicorn == 19.9.0
opencv-python == 4.5.2.52
tensorflow>=19.0
